Upgrade `babel-plugin-flow-prop-types` (#1045)
* Fix build

* Latest version fixes Header

* Fix propTypes from react-native leaking to web context

* Update todo notice
