+++ website/package.json
@@ -9,6 +9,7 @@
     "babel-eslint": "7.0.0",
     "babel-jest": "16.0.0",
     "babel-loader": "6.2.5",
+    "babel-plugin-flow-react-proptypes": "grabbou/babel-plugin-flow-react-proptypes#release",
     "babel-preset-react-app": "^1.0.0",
     "case-sensitive-paths-webpack-plugin": "1.1.4",
     "chalk": "1.1.3",
